test: dtoss 5104 eligibility flag check add/amend participant record (#844)
* test: fix appsettings-example reference

* test: verify eligibility flag is set to true for add participant

* test: amend record with eligibility as true

* test: chore removed not required imports

* test: chore correct test description for amend to match add

* test: chore match naming convetion structure for features

* test: isolate common steps

* test: fix spell Exception_Management

* test: chore cleaner feature and step implementation
